Added EOF and fixed tree algorithm
An EOF character '\0' is added to be also encoded
in the Huffman Tree. This will be extremely useful
for decoding binary files later as this will allow
the decoder to know where the end of the file is.

The Huffman tree algorithm was also fixed. Apparently,
there were cases where after fetching two character
nodes, the  new root was not properly updated, which
causes frequent characters to have longer binary
representation than necessary.

Signed-off-by: PaimonCodes <cryptsGuard@gmail.com>